Agentic-Retrieval-with-Image-Serving/requirements.txt
azure-search-documents
azure-core
azure-identity
python-dotenv
requests
